api/requirements.txt
asgiref==3.6.0
#backports.zoneinfo==0.2.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
class-registry==2.1.2
colorama==0.4.3
config==0.5.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-api-logger==1.1.12
drf-yasg==1.21.5
filters==1.3.2
filters-django==1.0.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.1
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.2
whitenoise==6.4.0
xlrd==2.0.1
